[chip-tool] Add src/app/tests/suites/include/PICSChecker.h and use it into examples/chip-tool and examples/placeholder
